+++ agents/src/cli.ts
@@ -32,23 +32,34 @@ const runWorker = async (args: CliArgs) => {
   }
 
   process.once('SIGINT', async () => {
+    logger.info('SIGINT received in CLI');
     // allow C-c C-c for force interrupt
     process.once('SIGINT', () => {
-      logger.info('worker closed forcefully');
+      logger.info('worker closed forcefully due to SIGINT.');
       process.exit(130); // SIGINT exit code
     });
     if (args.production) {
       await worker.drain();
     }
     await worker.close();
-    logger.info('worker closed');
+    logger.info('worker closed due to SIGINT.');
     process.exit(130); // SIGINT exit code
+  });
+
+  process.once('SIGTERM', async () => {
+    logger.info('SIGTERM received in CLI.');
+    if (args.production) {
+      await worker.drain();
+    }
+    await worker.close();
+    logger.info('worker closed due to SIGTERM.');
+    process.exit(143); // SIGTERM exit code
   });
 
   try {
     await worker.run();
   } catch {
-    logger.fatal('worker failed');
+    logger.fatal('closing worker due to error.');
     process.exit(1);
   }
 };
@@ -90,6 +101,11 @@ export const runApp = (opts: WorkerOptions) => {
         'LIVEKIT_API_SECRET',
       ),
     )
+    .addOption(
+      new Option('--worker-token <string>', 'Internal use only')
+        .env('LIVEKIT_WORKER_TOKEN')
+        .hideHelp(),
+    )
     .action(() => {
       if (
         // do not run CLI if origin file is agents/ipc/job_main.js
@@ -109,6 +125,7 @@ export const runApp = (opts: WorkerOptions) => {
       opts.apiKey = options.apiKey || opts.apiKey;
       opts.apiSecret = options.apiSecret || opts.apiSecret;
       opts.logLevel = options.logLevel || opts.logLevel;
+      opts.workerToken = options.workerToken || opts.workerToken;
       runWorker({
         opts,
         production: true,
@@ -131,6 +148,7 @@ export const runApp = (opts: WorkerOptions) => {
       opts.apiKey = options.apiKey || opts.apiKey;
       opts.apiSecret = options.apiSecret || opts.apiSecret;
       opts.logLevel = options.logLevel || opts.logLevel;
+      opts.workerToken = options.workerToken || opts.workerToken;
       runWorker({
         opts,
         production: false,
@@ -155,6 +173,7 @@ export const runApp = (opts: WorkerOptions) => {
       opts.apiKey = options.apiKey || opts.apiKey;
       opts.apiSecret = options.apiSecret || opts.apiSecret;
       opts.logLevel = options.logLevel || opts.logLevel;
+      opts.workerToken = options.workerToken || opts.workerToken;
       runWorker({
         opts,
         production: false,