+++ agents/src/multimodal/multimodal_agent.ts
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 import type {
   LocalTrackPublication,
+  NoiseCancellationOptions,
   RemoteAudioTrack,
   RemoteParticipant,
   RemoteTrack,
@@ -19,6 +20,11 @@ import {
 } from '@livekit/rtc-node';
 import { EventEmitter } from 'node:events';
 import { AudioByteStream } from '../audio.js';
+import {
+  ATTRIBUTE_TRANSCRIPTION_FINAL,
+  ATTRIBUTE_TRANSCRIPTION_TRACK_ID,
+  TOPIC_TRANSCRIPTION,
+} from '../constants.js';
 import * as llm from '../llm/index.js';
 import { log } from '../log.js';
 import type { MultimodalLLMMetrics } from '../metrics/base.js';
@@ -72,17 +78,20 @@ export class MultimodalAgent extends EventEmitter {
     chatCtx,
     fncCtx,
     maxTextResponseRetries = 5,
+    noiseCancellation,
   }: {
     model: RealtimeModel;
     chatCtx?: llm.ChatContext;
     fncCtx?: llm.FunctionContext;
     maxTextResponseRetries?: number;
+    noiseCancellation?: NoiseCancellationOptions;
   }) {
     super();
     this.model = model;
     this.#chatCtx = chatCtx;
     this.#fncCtx = fncCtx;
     this.#maxTextResponseRetries = maxTextResponseRetries;
+    this.#noiseCancellation = noiseCancellation;
   }
 
   #participant: RemoteParticipant | string | null = null;
@@ -95,6 +104,7 @@ export class MultimodalAgent extends EventEmitter {
   #session: RealtimeSession | null = null;
   #fncCtx: llm.FunctionContext | undefined = undefined;
   #chatCtx: llm.ChatContext | undefined = undefined;
+  #noiseCancellation: NoiseCancellationOptions | undefined = undefined;
 
   #_started: boolean = false;
   #_pendingFunctionCalls: Set<string> = new Set();
@@ -246,8 +256,8 @@ export class MultimodalAgent extends EventEmitter {
         if (message.contentType === 'text') return;
 
         const synchronizer = new TextAudioSynchronizer(defaultTextSyncOptions);
-        synchronizer.on('textUpdated', (text) => {
-          this.#publishTranscription(
+        synchronizer.on('textUpdated', async (text) => {
+          await this.#publishTranscription(
             this.room!.localParticipant!.identity!,
             this.#getLocalTrackSid()!,
             text.text,
@@ -297,25 +307,31 @@ export class MultimodalAgent extends EventEmitter {
       });
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      this.#session.on('input_speech_committed', (ev: any) => {
+      this.#session.on('input_speech_committed', async (ev: any) => {
         // openai.realtime.InputSpeechCommittedEvent
         const participantIdentity = this.linkedParticipant?.identity;
         const trackSid = this.subscribedTrack?.sid;
         if (participantIdentity && trackSid) {
-          this.#publishTranscription(participantIdentity, trackSid, '…', false, ev.itemId);
+          await this.#publishTranscription(participantIdentity, trackSid, '…', false, ev.itemId);
         } else {
           this.#logger.error('Participant or track not set');
         }
       });
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      this.#session.on('input_speech_transcription_completed', (ev: any) => {
+      this.#session.on('input_speech_transcription_completed', async (ev: any) => {
         // openai.realtime.InputSpeechTranscriptionCompletedEvent
         const transcription = ev.transcript;
         const participantIdentity = this.linkedParticipant?.identity;
         const trackSid = this.subscribedTrack?.sid;
         if (participantIdentity && trackSid) {
-          this.#publishTranscription(participantIdentity, trackSid, transcription, true, ev.itemId);
+          await this.#publishTranscription(
+            participantIdentity,
+            trackSid,
+            transcription,
+            true,
+            ev.itemId,
+          );
         } else {
           this.#logger.error('Participant or track not set');
         }
@@ -327,7 +343,7 @@ export class MultimodalAgent extends EventEmitter {
         this.#logger.child({ transcription }).debug('committed user speech');
       });
 
-      this.#session.on('input_speech_started', (ev: any) => {
+      this.#session.on('input_speech_started', async (ev: any) => {
         this.emit('user_started_speaking');
         if (this.#playingHandle && !this.#playingHandle.done) {
           this.#playingHandle.interrupt();
@@ -344,7 +360,7 @@ export class MultimodalAgent extends EventEmitter {
         const participantIdentity = this.linkedParticipant?.identity;
         const trackSid = this.subscribedTrack?.sid;
         if (participantIdentity && trackSid) {
-          this.#publishTranscription(participantIdentity, trackSid, '…', false, ev.itemId);
+          await this.#publishTranscription(participantIdentity, trackSid, '…', false, ev.itemId);
         }
       });
 
@@ -454,9 +470,12 @@ export class MultimodalAgent extends EventEmitter {
     this.subscribedTrack = track;
 
     this.readMicroTask = new Promise<void>((resolve, reject) => {
-      readAudioStreamTask(new AudioStream(track, this.model.sampleRate, this.model.numChannels))
-        .then(resolve)
-        .catch(reject);
+      const audioStreamOptions = {
+        sampleRate: this.model.sampleRate,
+        numChannels: this.model.numChannels,
+        ...(this.#noiseCancellation ? { noiseCancellation: this.#noiseCancellation } : {}),
+      };
+      readAudioStreamTask(new AudioStream(track, audioStreamOptions)).then(resolve).catch(reject);
     });
   }
 
@@ -467,13 +486,13 @@ export class MultimodalAgent extends EventEmitter {
     return this.#localTrackSid;
   }
 
-  #publishTranscription(
+  async #publishTranscription(
     participantIdentity: string,
     trackSid: string,
     text: string,
     isFinal: boolean,
     id: string,
-  ): void {
+  ): Promise<void> {
     this.#logger.debug(
       `Publishing transcription ${participantIdentity} ${trackSid} ${text} ${isFinal} ${id}`,
     );
@@ -496,6 +515,17 @@ export class MultimodalAgent extends EventEmitter {
         },
       ],
     });
+
+    const stream = await this.room.localParticipant.streamText({
+      topic: TOPIC_TRANSCRIPTION,
+      senderIdentity: participantIdentity,
+      attributes: {
+        [ATTRIBUTE_TRANSCRIPTION_TRACK_ID]: trackSid,
+        [ATTRIBUTE_TRANSCRIPTION_FINAL]: isFinal.toString(),
+      },
+    });
+    await stream.write(text);
+    await stream.close();
   }
 
   #updateState() {