+++ plugins/openai/src/realtime/realtime_model.ts
@@ -399,7 +399,7 @@ export class RealtimeModel extends multimodal.RealtimeModel {
   }) {
     super();
 
-    if (apiKey === '') {
+    if (apiKey === '' && !(isAzure && entraToken)) {
       throw new Error(
         'OpenAI API key is required, either using the argument or by setting the OPENAI_API_KEY environmental variable',
       );
@@ -599,6 +599,7 @@ export class RealtimeSession extends multimodal.RealtimeSession {
     temperature = this.#opts.temperature,
     maxResponseOutputTokens = this.#opts.maxResponseOutputTokens,
     toolChoice = 'auto',
+    selectedTools = Object.keys(this.#fncCtx || {}),
   }: {
     modalities: ['text', 'audio'] | ['text'];
     instructions?: string;
@@ -610,6 +611,7 @@ export class RealtimeSession extends multimodal.RealtimeSession {
     temperature?: number;
     maxResponseOutputTokens?: number;
     toolChoice?: api_proto.ToolChoice;
+    selectedTools?: string[];
   }) {
     this.#opts = {
       modalities,
@@ -630,16 +632,18 @@ export class RealtimeSession extends multimodal.RealtimeSession {
     };
 
     const tools = this.#fncCtx
-      ? Object.entries(this.#fncCtx).map(([name, func]) => ({
-          type: 'function' as const,
-          name,
-          description: func.description,
-          parameters:
-            // don't format parameters if they are raw openai params
-            func.parameters.type == ('object' as const)
-              ? func.parameters
-              : llm.oaiParams(func.parameters),
-        }))
+      ? Object.entries(this.#fncCtx)
+          .filter(([name]) => selectedTools.includes(name))
+          .map(([name, func]) => ({
+            type: 'function' as const,
+            name,
+            description: func.description,
+            parameters:
+              // don't format parameters if they are raw openai params
+              func.parameters.type == ('object' as const)
+                ? func.parameters
+                : llm.oaiParams(func.parameters),
+          }))
       : [];
 
     const sessionUpdateEvent: api_proto.SessionUpdateEvent = {