+++ plugins/silero/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @livekit/agents-plugin-silero
 
+## 0.5.6
+
+### Patch Changes
+
+- cleanup resources with onnx runtime - [#377](https://github.com/livekit/agents-js/pull/377) ([@Shubhrakanti](https://github.com/Shubhrakanti))
+
+- Updated dependencies [[`ae508d76d6b521218f8320f8dba1b1d0fdad79a2`](https://github.com/livekit/agents-js/commit/ae508d76d6b521218f8320f8dba1b1d0fdad79a2), [`e1dd8f86660b4f96585833dc896b719dd3ad54b2`](https://github.com/livekit/agents-js/commit/e1dd8f86660b4f96585833dc896b719dd3ad54b2)]:
+  - @livekit/agents@0.7.5
+
+## 0.5.5
+
+### Patch Changes
+
+- fix frames having erroneous bits of audio at start - [#310](https://github.com/livekit/agents-js/pull/310) ([@nbsp](https://github.com/nbsp))
+
+- Updated dependencies [[`724c02bb7a91c27d6c8daf961842fb9f0934770c`](https://github.com/livekit/agents-js/commit/724c02bb7a91c27d6c8daf961842fb9f0934770c), [`7398cffad62b17c79b5fe2f0ca4e99e548560367`](https://github.com/livekit/agents-js/commit/7398cffad62b17c79b5fe2f0ca4e99e548560367), [`6ed0c90d1bab013854c416768b10ef96f3227d68`](https://github.com/livekit/agents-js/commit/6ed0c90d1bab013854c416768b10ef96f3227d68), [`33c241960f0e8f325f534d2406f42148a4486b5a`](https://github.com/livekit/agents-js/commit/33c241960f0e8f325f534d2406f42148a4486b5a)]:
+  - @livekit/agents@0.7.1
+
 ## 0.5.4
 
 ### Patch Changes